chore: Make the windows code compatible with TCHAR.
This adds support for multi-byte builds. Users building qTox in multi-byte
instead of unicode mode would otherwise get compiler errors. This is one step
in the direction of making qTox a library with the UI being just one application
of it: it gives library users a choice of how they build it.